Refactors a bit
